Исправляет функцию update у классов
